requirements.txt
appdirs==1.4.4
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
idna==3.4
lxml==4.9.2
python-dotenv==0.21.1
requests==2.28.2
requests-cache==0.9.8
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.14
